Moves cardList and getSpyCard to helpers/
